Fix foreign-key type mismatch that broke a clean MySQL migrate
Foreign-key columns referencing bigint id() primary keys (users, product,
invoice, quationform) were declared as integer, not unsignedBigInteger.
On a strict MySQL server a fresh `php artisan migrate` failed at the jobs
table with "Referencing column 'created_by' and referenced column 'id'
are incompatible" (errno 3780). SQLite doesn't enforce FK column-type
compatibility, so the test suite never caught it; existing live databases
work only because they were built incrementally.

Converts every integer FK-source column with a foreign() constraint to
unsignedBigInteger across 16 migrations (created_by, assigned_to,
decided_by, paid_by, marked_by, completed_by, uploaded_by, user_id,
product_id, invoice_id, quotation_id). Verified: a from-scratch migrate
of all 67 migrations now completes with 0 failures and the demo seeder
runs clean on it. Full suite still 495 passing.

Regenerates database/demo/demo-data.sql from the clean-migrated schema so
the shipped snapshot matches the corrected column types.

+++ database/migrations/2026_07_18_000007_create_jobs_table.php
@@ -20,12 +20,12 @@ return new class extends Migration
             // type MySQL refuses as FK-incompatible with the real users table (only
             // caught by testing against live MySQL; sqlite's weak typing let this pass
             // silently in the automated test suite).
-            $table->integer('created_by');
-            $table->integer('assigned_to')->nullable();
+            $table->unsignedBigInteger('created_by');
+            $table->unsignedBigInteger('assigned_to')->nullable();
             $table->enum('priority', ['low', 'medium', 'high', 'urgent'])->default('medium');
             $table->date('due_date')->nullable();
             $table->enum('status', ['pending_approval', 'assigned', 'in_progress', 'on_hold', 'completed', 'rejected']);
-            $table->integer('decided_by')->nullable();
+            $table->unsignedBigInteger('decided_by')->nullable();
             $table->timestamp('decided_at')->nullable();
             $table->text('rejection_reason')->nullable();
             $table->text('on_hold_reason')->nullable();
